requirements.txt
streamlit
numpy==1.26.4
torch
torchvision
facenet-pytorch
pillow
joblib
